add domain label
