+++ pyproject.toml
@@ -17,7 +17,7 @@ prometheus-client = "0.14.0"
 confluent-kafka = "1.8.2"
 flashbots = {git = "https://github.com/lidofinance/web3-flashbots.git", rev = "v1.1.0"}
 timeout-decorator = "^0.5.0"
-web3-multi-provider = "^0.4.0"
+web3-multi-provider = "^0.4.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.5"
