+++ scripts/utils/metrics.py
@@ -69,6 +69,6 @@ ETH_RPC_REQUESTS_DURATION = Histogram(
 ETH_RPC_REQUESTS = Counter(
     'eth_rpc_requests',
     'Total count of requests to ETH1 RPC',
-    ['method', 'code'],
+    ['method', 'code', 'domain'],
     namespace=PREFIX
 )
